requirements.txt
asgiref==3.9.0
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
diff-match-patch==20241021
Django==4.2.23
django-cors-headers==4.3.1
django-filter==23.4
django-import-export==4.3.8
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pillow==11.3.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
